refactor: dedupe #423's test infrastructure per /simplify findings
Extracted EvalRunnerTestComposition.BuildServices() -- the sibling
EvalRunnerValidateOnBuildTests and the new
MetricSpecParameterKeyValidationTests were independently building the
identical six-call DI composition sequence, both explicitly commenting
that they mirrored the other rather than sharing it.

Replaced the new test's hand-rolled walk-up-from-AppContext.BaseDirectory
directory search with RepoRoot.Combine(...) -- Tests.Common.RepoRoot
already centralizes this exact walk after it replaced five duplicate
copies of the same pattern, one of which broke test discovery in git
worktrees (#293). The new test had copied the sibling
SeedDatasetSmokeTests' pre-existing (separately out-of-scope) copy of
the same pattern instead of the shared fix.

Replaced IEvalMetric's own hand-rolled empty HashSet<string> field with
ImmutableHashSet<string>.Empty, matching the BCL-cached-singleton
convention already used elsewhere in the codebase for exactly this.

Filed #437 for a genuine altitude-angle finding: RecognizedParameterKeys
has zero runtime consumers today, so only the repo's own checked-in
eval-datasets are protected -- a template consumer's own dataset gets
no benefit. Also identified a structurally identical gap in
EvalCase.InvocationOverrides (IAgentInvoker). Both are real
enhancements beyond #423's originating issue's literal scope, not
defects in what shipped -- tracked as a follow-up rather than expanding
this PR further.
