+++ src/Content/Tests/Presentation.EvalRunner.Tests/Composition/EvalRunnerValidateOnBuildTests.cs
@@ -1,10 +1,6 @@
 using Application.AI.Common.Evaluation.Interfaces;
-using Domain.Common.Config;
 using FluentAssertions;
-using Infrastructure.AI.Evaluation;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Presentation.Common.Extensions;
 using Xunit;
 
 namespace Presentation.EvalRunner.Tests.Composition;
@@ -19,27 +15,10 @@ namespace Presentation.EvalRunner.Tests.Composition;
 /// </summary>
 public sealed class EvalRunnerValidateOnBuildTests
 {
-    private static ServiceCollection BuildEvalRunnerServices()
-    {
-        var configuration = new ConfigurationBuilder()
-            .AddInMemoryCollection(new Dictionary<string, string?>())
-            .Build();
-
-        var services = new ServiceCollection();
-        services.AddLogging();
-        services.RegisterConfigSections(configuration);
-        var appConfig = configuration.GetSection("AppConfig").Get<AppConfig>() ?? new AppConfig();
-        services.BuildGlobalSolutionServices(appConfig, includeHealthChecksUI: false);
-
-        // Mirror Program.cs: the eval host opts into the framework AFTER the shared root.
-        services.AddEvaluationDependencies();
-        return services;
-    }
-
     [Fact]
     public void EvalRunnerComposition_BuildsWithValidateOnBuild()
     {
-        var services = BuildEvalRunnerServices();
+        var services = EvalRunnerTestComposition.BuildServices();
 
         var exception = Record.Exception(() =>
         {
@@ -56,7 +35,7 @@ public sealed class EvalRunnerValidateOnBuildTests
     [Fact]
     public void EvalRunnerComposition_ResolvesRealEvalRunner_NotTheNotConfiguredDefault()
     {
-        var services = BuildEvalRunnerServices();
+        var services = EvalRunnerTestComposition.BuildServices();
 
         // No validation flags needed here — this asserts resolution only; the sibling
         // BuildsWithValidateOnBuild test pins the validation policy for this same graph.
