+++ src/Content/Tests/Presentation.EvalRunner.Tests/Presentation.EvalRunner.Tests.csproj
@@ -19,6 +19,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../../Presentation/Presentation.EvalRunner/Presentation.EvalRunner.csproj" />
+    <ProjectReference Include="../Tests.Common/Tests.Common.csproj" />
   </ItemGroup>
 
 </Project>
